[Fixed]
- 자동 로그인 버그 수정
- 인트로 페이지 추가
