+++ src/pages/create/CreateOne.tsx
@@ -33,7 +33,7 @@ export function CreateOne() {
     setIsOpen(false);
   };
   const handleConfirm = () => {
-    navigate('/');
+    navigate('/main');
   };
   const onSubmit = (data: FormValues) => {
     console.log(data);
