+++ src/pages/index.ts
@@ -8,3 +8,4 @@ export * from './signup';
 export * from './detail';
 export * from './review';
 export * from './category';
+export * from './intro';
