+++ src/pages/signup/SignupComplete.tsx
@@ -6,7 +6,7 @@ export function SignupComplete() {
     <FullPageInfo
       icons={[checkCirclePrimary, '회원가입 완료']}
       text={['축하드립니다', '가입 완료!']}
-      route={() => '/'}
+      route={() => '/main'}
       hasDetailedDescription={false}
     />
   );
