+++ src/pages/signup/Verification.tsx
@@ -34,7 +34,7 @@ export function Verification() {
 
   // 모달창 확인 버튼 클릭 시
   const handleConfirm = () => {
-    navigate('/');
+    navigate('/main');
   };
 
   const navigate = useNavigate();
