+++ src/pages/user/MyRecipes.tsx
@@ -82,7 +82,7 @@ export function MyRecipes() {
   const handleConfirm = () => {
     sessionStorage.clear();
     localStorage.removeItem('pocketbase_auth');
-    navigate('/');
+    navigate('/main');
   };
 
   const handleLogout = () => {
